+++ public/results.php
@@ -347,7 +347,7 @@ if (!empty($results[$p]) && count($results[$p]) > 0) {
               <span class="highlight"><?php echo $file['path_parent']; ?></span>
           </td>
         <td class="text-nowrap highlight" style="font-weight:bold;color:#D20915;"><?php echo formatBytes($file['filesize']); ?></td>
-        <td width="8%" class="highlight"><div class="text-right percent" style="width:<?php echo number_format(($file['filesize'] / $total_size) * 100, 2); ?>%;"></div> <span style="color:gray;"><small><?php echo number_format(($file['filesize'] / $total_size) * 100, 2); ?>%</small></span></td>
+        <td width="8%" class="highlight"><div class="text-right percent" style="width:<?php echo ($total_size > 0) ? number_format(($file['filesize'] / $total_size) * 100, 2) : number_format(0, 2); ?>%;"></div> <span style="color:gray;"><small><?php echo ($total_size > 0) ? number_format(($file['filesize'] / $total_size) * 100, 2) : number_format(0, 2); ?>%</small></span></td>
         <?php if ($_GET['doctype'] == 'directory' || $_GET['doctype'] == '') { ?>
         <!-- show comparison file size -->
         <?php if ($show_change_percent) { $filesize_change = 0; ?>